refactor(channels): 将 WebuiModalContent 内联到 channels 页面并移除独立文件
